# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Template API Examples",
        "title": "Template API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "paths": {
        "/liveness": {
            "get": {
                "description": "Checks if the application is alive and running (heartbeat check)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Get liveness health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_healthz.livenessStats"
                        }
                    }
                }
            }
        },
        "/readiness": {
            "get": {
                "description": "Checks if the application is ready to accept traffic (dependencies ready)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Get readiness health status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_healthz.readinessStats"
                        }
                    }
                }
            }
        },
        "/v1/xxxxx": {
            "get": {
                "description": "Get xxxxx records with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "xxxxx"
                ],
                "summary": "Get Xxxxx with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.SwaggerGetXxxxxPagingResponse"
                        }
                    },
                    "400": {
                        "description": "type: bind_data_error, validation_error",
                        "schema": {
                            "$ref": "#/definitions/xres.BadRequestResponse"
                        }
                    },
                    "422": {
                        "description": "type: operation_failed",
                        "schema": {
                            "$ref": "#/definitions/xres.UnprocessableEntityResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new xxxxx record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "xxxxx"
                ],
                "summary": "Create a new Xxxxx",
                "parameters": [
                    {
                        "description": "Create xxxxx request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.PostXxxxxRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.SwaggerPostXxxxxResponse"
                        }
                    },
                    "400": {
                        "description": "type: bind_data_error, validation_error",
                        "schema": {
                            "$ref": "#/definitions/xres.BadRequestResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/xres.UnprocessableEntityResponse"
                        }
                    }
                }
            }
        },
        "/v1/xxxxx/{id}": {
            "get": {
                "description": "Get a specific xxxxx record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "xxxxx"
                ],
                "summary": "Get Xxxxx by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Xxxxx ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.SwaggerGetXxxxxByIDResponse"
                        }
                    },
                    "422": {
                        "description": "type: operation_failed",
                        "schema": {
                            "$ref": "#/definitions/xres.UnprocessableEntityResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_healthz.livenessStats": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Liveness healthy"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "type": "string",
                    "example": "liveness"
                }
            }
        },
        "internal_healthz.readinessStats": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Readiness healthy"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "type": "string",
                    "example": "readiness"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.GetXxxxxByIDResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.GetXxxxxPagingResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.PostXxxxxRequest": {
            "type": "object",
            "required": [
                "code",
                "name_en",
                "name_th"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "name_th": {
                    "type": "string"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.PostXxxxxResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.SwaggerGetXxxxxByIDResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.GetXxxxxByIDResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Successful."
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.SwaggerGetXxxxxPagingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.GetXxxxxPagingResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successful."
                },
                "meta": {
                    "$ref": "#/definitions/xres.PagingMeta"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "template-api-golang_internal_api_xxxxxs_models.SwaggerPostXxxxxResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/template-api-golang_internal_api_xxxxxs_models.PostXxxxxResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Created."
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "type": {
                    "type": "string",
                    "example": "created"
                }
            }
        },
        "xres.BadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad request."
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "type": {
                    "type": "string",
                    "example": "bad_request"
                }
            }
        },
        "xres.PagingMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "offset": {
                    "type": "integer",
                    "example": 0
                },
                "total_items": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "xres.UnprocessableEntityResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Unprocessable entity."
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "type": {
                    "type": "string",
                    "example": "unprocessable_entity"
                }
            }
        }
    }
}